requirements.txt
pillow
tensorflow
Flask
gunicorn
numpy